refactor(orb-live): A3 - extract buildLiveSystemInstruction (VTID-02900)
Lifts the system instruction builder + its private helpers into a new module:

  services/gateway/src/orb/live/instruction/live-system-instruction.ts (NEW)
    - TemporalBucket (type)
    - describeTimeSince  (exported, was already exported for A0.2)
    - describeRoute      (exported, was private)
    - buildTemporalJourneyContextSection (private to the new module)
    - buildLiveSystemInstruction (exported)

orb-live.ts re-exports these so external callers (including the
characterization test suite) continue to resolve the same symbols
unchanged.

Helper that stays in orb-live.ts:
  - buildNavigatorPolicySection — exported so the lifted instruction
    builder can call back across the boundary. Pure function (lang → string),
    no circular-import risk.

Verification:
  - 133 characterization tests pass, 9 snapshots match byte-for-byte.
  - tsc --noEmit: zero new errors.

This unblocks B0 work — the instruction builder is now the single
insertion point for B0a (Situational Awareness Core), B0d (Continuation
Contract) and B0e (Feature Discovery) per the approved plan.

Part of Track A, step A3.
